+++ package.json
@@ -36,6 +36,7 @@
         "@azure/ms-rest-nodeauth": "^3.0.6",
         "@azure/storage-blob": "^12.2.1",
         "async": "^3.2.0",
+        "azure-storage": "^2.10.3",
         "bcrypt-nodejs": "^0.0.3",
         "bluebird": "^3.7.2",
         "body-parser": "^1.19.0",
@@ -71,6 +72,7 @@
         "@types/errorhandler": "^0.0.32",
         "@types/eslint": "^6.8.1",
         "@types/express": "^4.17.11",
+        "@types/express-fileupload": "^1.1.6",
         "@types/express-flash": "0.0.1",
         "@types/express-session": "1.15.16",
         "@types/jest": "^24.9.1",
@@ -94,6 +96,7 @@
         "chai": "^4.3.0",
         "concurrently": "^5.3.0",
         "eslint": "^6.8.0",
+        "express-fileupload": "^1.2.1",
         "fx": "^20.0.2",
         "jest": "^26.6.3",
         "node-sass": "^4.14.1",
